requirements.txt
flask>=3.0.0
line-bot-sdk>=3.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
gunicorn>=21.0.0
redis>=5.0.0
google-genai>=0.3.0
Pillow>=10.0.0
